Don't create buddies that are hidden or blocked.
Described in https://github.com/EionRobb/purple-googlechat/issues/37
there is an issues where unwanted buddies show up in the buddy list, and
if you remove them, they reappear the next time you restart pidgin, or
reconnect to Google Chat.

This is because the API we are using shows every DM session we have ever
had, and while we are being told which of them has been hidden, or
blocked, up until now the plugin has ignored those flags.

This fixes that, only bothering to add buddies that are neither hidden
nor blocked.

This finally allows someone to remove a buddy (which internally to the
plug hides the buddy), and have them stay gone.